// Doctor portal/src/app/models/doctor.model.ts
// daoctor obj in patient collections
export interface Patient_doctor{
    name?: string;
    mbbs?: string;
    email?: string;
    docID?: string;
    mobileNumber?: string;
    nic?: string;
    specialty?: string;
    status?: string;
}

export interface Doctor{
    id?: string;
    name?: string;
    mbbs?: string;
    email?: string;
    mobileNumber?: string;
    nic?: string;
    specialty?: string;
    gender?: string;
    uid?: string;
    bio?: string;
    proPic?: string;
}